chore: 🤖 api production install

+++ packages/leaa-api/deploy.sh
@@ -60,7 +60,13 @@ platform_test() {
 }
 
 platform_test_docker() {
-  cp -f ./.env.production ${__DEPLOY__}
+  if [ -f "${__DEPLOY__}/.env.production" ]; then
+      # shellcheck disable=SC2028
+      echo '\n✨  Already .env.production, Do not copy.\n'
+    else
+      cp -f ./.env.production ${__DEPLOY__}
+  fi
+
   cp -f ./docker-compose.yml ${__DEPLOY__}
   cd ${__DEPLOY__} || exit
